Maps not storing none can be annoying
This is the second time in just this project where using MAP! has bitten
me.

When creating the INTO node, if no type is specified set TYPE to FALSE.
This is better than leaving it unset (by setting it to NONE) because
there is some weirdness with binding if the word 'TYPE is not in the
context. (I *thought* I had everything copied so it wouldn't be an
issue, but for some reason I don't have that right. Much easier to just
have 'TYPE always there.)